+++ crates/client/src/config.rs
@@ -5,6 +5,7 @@ use std::env;
 use std::sync::{Arc, Mutex};
 use trust0_common::client::replshell_io::{ReplShellInputReader, ReplShellOutputWriter};
 use trust0_common::crypto::file::{load_certificates, load_private_key};
+use trust0_common::crypto::tls::danger;
 use trust0_common::distro::AppInstallFile;
 use trust0_common::error::AppError;
 use trust0_common::file;
@@ -215,74 +216,17 @@ fn spawn_line_reader(reader: &dyn ReplShellInputReader) {
 #[cfg(test)]
 fn spawn_line_reader(_reader: &dyn ReplShellInputReader) {}
 
-mod danger {
-    use pki_types::{CertificateDer, ServerName, UnixTime};
-    use rustls::client::danger::HandshakeSignatureValid;
-    use rustls::crypto::{verify_tls12_signature, verify_tls13_signature};
-    use rustls::DigitallySignedStruct;
-
-    #[derive(Debug)]
-    pub struct NoCertificateVerification {}
-
-    impl rustls::client::danger::ServerCertVerifier for NoCertificateVerification {
-        fn verify_server_cert(
-            &self,
-            _end_entity: &CertificateDer<'_>,
-            _intermediates: &[CertificateDer<'_>],
-            _server_name: &ServerName<'_>,
-            _ocsp: &[u8],
-            _now: UnixTime,
-        ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
-            Ok(rustls::client::danger::ServerCertVerified::assertion())
-        }
-
-        fn verify_tls12_signature(
-            &self,
-            message: &[u8],
-            cert: &CertificateDer<'_>,
-            dss: &DigitallySignedStruct,
-        ) -> Result<HandshakeSignatureValid, rustls::Error> {
-            verify_tls12_signature(
-                message,
-                cert,
-                dss,
-                &rustls::crypto::ring::default_provider().signature_verification_algorithms,
-            )
-        }
-
-        fn verify_tls13_signature(
-            &self,
-            message: &[u8],
-            cert: &CertificateDer<'_>,
-            dss: &DigitallySignedStruct,
-        ) -> Result<HandshakeSignatureValid, rustls::Error> {
-            verify_tls13_signature(
-                message,
-                cert,
-                dss,
-                &rustls::crypto::ring::default_provider().signature_verification_algorithms,
-            )
-        }
-
-        fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-            rustls::crypto::ring::default_provider()
-                .signature_verification_algorithms
-                .supported_schemes()
-        }
-    }
-}
-
 /// Unit tests
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use crate::config::danger::NoCertificateVerification;
     use crate::console::ShellInputReader;
     use pki_types::{ServerName, UnixTime};
     use rustls::client::danger::ServerCertVerifier;
     use std::env;
     use std::path::PathBuf;
     use std::sync::atomic::{AtomicBool, Ordering};
+    use trust0_common::crypto::tls::danger::NoCertificateVerification;
     use trust0_common::testutils::TEST_MUTEX;
 
     const CONFIG_FILE_PATHPARTS: [&str; 3] =
